jpa, annotations, color, rgba, fixes

+++ src/main/java/de/telran/springdemo/service/GreetingService.java
@@ -1,11 +1,11 @@
 package de.telran.springdemo.service;
 
-import de.telran.springdemo.model.Greeting;
+import de.telran.springdemo.entity.Greeting;
 
 public interface GreetingService {
     long create( Greeting greeting);
 
     Greeting get(long id);
 
-    void update(int id, int count);
+    void update(long id, int count);
 }
